Factor our common service methods
